pyproject.toml
[project]
    name = "avatar-back"
    version = "0.1.0"
    description = "Leveraging the capabilities of LLMs to improve day-to-day tasks."
    authors = [
        { name = "Alex", email = "30626291+ipruning@users.noreply.github.com" },
    ]
    readme = "README.md"
    requires-python = ">= 3.12"
    dependencies = [
        "autocorrect-py>=2.9.0",
        "banks>=0.4.1",
        "dspy-ai>=2.4.12",
        "httpx[socks]>=0.27.0",
        "llama-index>=0.10.55",
        "openai>=1.35.14",
        "python-dotenv>=1.0.1",
        "sglang[openai]>=0.1.21",
        "tenacity>=8.5.0",
        "tqdm>=4.66.4",
        "anthropic>=0.31.2",
        "replicate>=0.28.0",
    ]

[build-system]
    requires      = ["hatchling"]
    build-backend = "hatchling.build"

[tool.rye]
    managed = true
    dev-dependencies = [
        "commitizen>=3.27.0",
        "coverage[toml]>=7.5.4",
        "debugpy>=1.8.2",
        "mypy>=1.10.1",
        "nox>=2024.4.15",
        "pre-commit>=3.7.1",
        "pytest-cov>=5.0.0",
        "pytest>=8.2.2",
        "ruff>=0.4.10",
    ]

[tool.hatch.metadata]
    allow-direct-references = true

[tool.hatch.build.targets.wheel]
    packages = ["src/tghl"]

[tool.ruff]
    # https://docs.astral.sh/ruff/configuration/
    # Exclude a variety of commonly ignored directories.
    exclude = [
        ".bzr",
        ".direnv",
        ".eggs",
        ".git",
        ".git-rewrite",
        ".hg",
        ".ipynb_checkpoints",
        ".mypy_cache",
        ".nox",
        ".pants.d",
        ".pyenv",
        ".pytest_cache",
        ".pytype",
        ".ruff_cache",
        ".svn",
        ".tox",
        ".venv",
        ".vscode",
        "__pypackages__",
        "_build",
        "buck-out",
        "build",
        "dist",
        "node_modules",
        "site-packages",
        "venv",
    ]

    # Same as Black.
    line-length  = 120
    indent-width = 4

[tool.ruff.lint]
    # Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
    # Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
    # McCabe complexity (`C901`) by default.
    select = ["E4", "E7", "E9", "F"]
    ignore = ["E501"]

    # Allow fix for all enabled rules (when `--fix`) is provided.
    fixable   = ["ALL"]
    unfixable = []

    # Allow unused variables when underscore-prefixed.
    dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
    # Like Black, use double quotes for strings.
    quote-style = "double"

    # Like Black, indent with spaces, rather than tabs.
    indent-style = "space"

    # Like Black, respect magic trailing commas.
    skip-magic-trailing-comma = false

    # Like Black, automatically detect the appropriate line ending.
    line-ending = "auto"

    # Enable auto-formatting of code examples in docstrings. Markdown,
    # reStructuredText code/literal blocks and doctests are all supported.
    #
    # This is currently disabled by default, but it is planned for this
    # to be opt-out in the future.
    docstring-code-format = false

    # Set the line length limit used when formatting code snippets in
    # docstrings.
    #
    # This only has an effect when the `docstring-code-format` setting is
    # enabled.
    docstring-code-line-length = "dynamic"

[tool.mypy]
    ignore_missing_imports = true
    strict                 = true

[tool.coverage.run]
    omit = [
        ".*",
        "*/site-packages/*",
        "*/__init__.py",
        "*/__main__.py",
        "*/tests",
        "*/excp.py",
    ]

[tool.coverage.report]
    fail_under   = 70
    show_missing = true

[tool.pytest.ini_options]
    python_files = ["test_*.py"]
    testpaths    = ["tests"]

[tool.pyright]
    venvPath = "."
    venv     = ".venv"
